requirements.txt
fastapi>=0.78.0
tortoise-orm>=0.19.0
uvicorn>=0.18.2
phonenumbers>=8.12.51